Show account switch toast when switching between tabs
